Remove get_tex_path module
My actual proposal for what to do about this is #673
